pyproject.toml
[project]
name = "gridstatus-forecast"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "gridstatus>=0.29.1",
    "gridstatusio>=0.15.1",
    "lightgbm>=4.6.0",
    "matplotlib>=3.9.4",
    "mlforecast>=1.0.2",
    "openpyxl>=3.1.5",
    "pandas>=2.3.3",
    "python-dotenv>=1.2.1",
    "pytz>=2025.2",
    "requests>=2.32.5",
    "seaborn>=0.13.2",
    "statsforecast>=2.0.3",
    "tqdm>=4.67.2",
    "window-ops>=0.0.15",
    "xgboost>=3.1.3",
]
